feat: Add basic Gateway for testing
I have develop a gateway with authentication for learning purposes, it
uses the controller - service architecture, with the following flow:

index.ts (entry point) -> middleware -> routes -> controllers ->
services -> repositories

The middleware is the one who checks if the client has a valid token
before access a route that is not public.

Routes will be the responsible of send each request to the proper
microservice.