fix: Fix auth-service validating token for Gateway
At this point, I have make the auth-service able to validate the token
for the gateway, but I have found that this can mean a lot of request
each time a client wants to go to a protected url, so I will change it
to be the gateway the ones who verify the tokens.

This is still a "working" version of the authentication proccess.