feat: untested algo ai in the backend
Now the ai container gets up and running but is not connected to anything yet. The idea is to connect it to the server-side-logic of the game and get another mode in the front to connect the ai websocket there.
The logic follows an algorithm to calculate the ball position after it bounces if it does. Also has some erratic movements programed into it to make it more human and clamping to prevent weird predictions and basing movements on them by using a sanity check.