feat: Remote private rooms working
Need more testing but now all pong services are up and running so the idea is to test the remote from two different computers. Also must check a lot of edge cases to see unexpected behaviour and fix it. So in theory it's ready for the new tournament mode. Also need to make the ai dumber.